Combine OpGraph and OperationGraph
